split up card_logic.py
